+++ charitable/includes/admin/views/settings/number.php
@@ -22,7 +22,7 @@ $max = isset( $view_args['max'] ) ? 'max="' . $view_args['max'] . '"' : '';
 <input type="number"
 	id="<?php printf( 'charitable_settings_%s', implode( '_', $view_args['key'] ) ); ?>"
 	name="<?php printf( 'charitable_settings[%s]', $view_args['name'] ); ?>"
-	value="<?php echo $value; ?>"
+	value="<?php echo esc_attr( $value ); ?>"
 	<?php echo $min; ?>
 	<?php echo $max; ?>
 	class="<?php echo esc_attr( $view_args['classes'] ); ?>"