+++ charitable/includes/emails/charitable-email-hooks.php
@@ -40,15 +40,11 @@ add_action( 'admin_init', array( Charitable_Emails::get_instance(), 'register_ad
  * @see Charitable_Email_Offline_Donation_Receipt::send_with_donation_id()
  * @see Charitable_Email_Offline_Donation_Notification::send_with_donation_id()
  */
+
 add_action( 'charitable_after_save_donation', array( 'Charitable_Email_Donation_Receipt', 'send_with_donation_id' ) );
 add_action( 'charitable_after_save_donation', array( 'Charitable_Email_New_Donation', 'send_with_donation_id' ) );
 add_action( 'charitable_after_save_donation', array( 'Charitable_Email_Offline_Donation_Receipt', 'send_with_donation_id' ) );
 add_action( 'charitable_after_save_donation', array( 'Charitable_Email_Offline_Donation_Notification', 'send_with_donation_id' ) );
-
-foreach ( charitable_get_approval_statuses() as $status ) {
-	add_action( $status . '_' . Charitable::DONATION_POST_TYPE, array( 'Charitable_Email_Donation_Receipt', 'send_with_donation_id' ) );
-	add_action( $status . '_' . Charitable::DONATION_POST_TYPE, array( 'Charitable_Email_New_Donation', 'send_with_donation_id' ) );
-}
 
 /**
  * Send the Campaign Ended email.